@@ -24,114 +24,73 @@
 #
 """Module covering information package structure validation and navigation."""
 import os
+from pathlib import Path
 from lxml import etree
 
 from eark_validator.const import NO_PATH, NOT_FILE, NOT_VALID_FILE
+from eark_validator.mets import MetsFiles, MetsFile
 from eark_validator.ipxml.namespaces import Namespaces
-from eark_validator.infopacks.manifest import Manifest
+from eark_validator.model import PackageDetails
+from eark_validator.model.package_details import InformationPackage
+from eark_validator.model.validation_report import Result
+from .package_handler import PackageHandler
 
-class PackageDetails:
+class InformationPackages:
 
-    def __init__(
-            self: str,
-            objid: str,
-            label: str,
-            type: str,
-            othertype: str,
-            contentinformationtype: str,
-            profile: str,
-            oaispackagetype: str,
-            ns: str):
-        self._objid = objid
-        self._label = label
-        self._type = type
-        self._othertype = othertype
-        self._contentinformationtype = contentinformationtype
-        self._profile = profile
-        self._oaispackagetype = oaispackagetype
-        self._ns = ns
-
-    @property
-    def objid(self) -> str:
-        return self._objid
-
-    @property
-    def label(self) -> str:
-        return self._label
-
-    @property
-    def type(self) -> str:
-        return self._type
-
-    @property
-    def othertype(self) -> str:
-        return self._othertype
-
-    @property
-    def contentinformationtype(self) -> str:
-        return self._contentinformationtype
-
-    @property
-    def profile(self) -> str:
-        return self._profile
-
-    @property
-    def oaispackagetype(self) -> str:
-        return self._oaispackagetype
-
-    @property
-    def namespaces(self) -> str:
-        return self._ns
-
-    @classmethod
-    def from_mets_file(cls, mets_file: str) -> 'PackageDetails':
-        if (not os.path.exists(mets_file)):
+    @staticmethod
+    def details_from_mets_file(mets_file: Path) -> PackageDetails:
+        if (not mets_file.exists()):
             raise FileNotFoundError(NO_PATH.format(mets_file))
-        if (not os.path.isfile(mets_file)):
+        if (not mets_file.is_file()):
             raise ValueError(NOT_FILE.format(mets_file))
         ns = {}
-        objid = label = ptype = othertype = contentinformationtype = profile = oaispackagetype = ''
+        label = othertype = contentinformationtype = oaispackagetype = ''
         try:
             parsed_mets = etree.iterparse(mets_file, events=['start', 'start-ns'])
             for event, element in parsed_mets:
                 if event == 'start-ns':
-                    prefix = element[0]
-                    ns_uri = element[1]
-                    ns[prefix] = ns_uri
+                    # Add namespace id to the dictionary
+                    ns[element[1]] = element[0]
                 if event == 'start':
                     if element.tag == Namespaces.METS.qualify('mets'):
-                        objid = element.get('OBJID', '')
                         label = element.get('LABEL', '')
-                        ptype = element.get('TYPE', '')
                         othertype = element.get(Namespaces.CSIP.qualify('OTHERTYPE'), '')
                         contentinformationtype = element.get(Namespaces.CSIP.qualify('CONTENTINFORMATIONTYPE'), '')
-                        profile = element.get('PROFILE', '')
                         oaispackagetype = element.find(Namespaces.METS.qualify('metsHdr')).get(Namespaces.CSIP.qualify('OAISPACKAGETYPE'), '')
-                    elif element.tag == Namespaces.METS.qualify('metsHdr'):
+                    else:
                         break
-        except etree.XMLSyntaxError:
+        except (etree.XMLSyntaxError, AttributeError):
             raise ValueError(NOT_VALID_FILE.format(mets_file, 'XML'))
-        return cls(objid, label, ptype, othertype, contentinformationtype, profile, oaispackagetype, ns)
+        return PackageDetails.model_validate({
+            'label': label,
+            'othertype': othertype,
+            'contentinformationtype': contentinformationtype,
+            'oaispackagetype': oaispackagetype
+        })
 
+    @staticmethod
+    def from_path(package_path: Path) -> InformationPackage:
+        if (not package_path.exists()):
+            raise FileNotFoundError(NO_PATH.format(package_path))
+        handler: PackageHandler = PackageHandler()
+        to_parse:Path = handler.prepare_package(package_path)
+        mets_path: Path = to_parse.joinpath('METS.xml')
+        if (not mets_path.is_file()):
+            raise ValueError('No METS file found in package')
+        mets: MetsFile = MetsFiles.from_file(to_parse.joinpath('METS.xml'))
+        return InformationPackage.model_validate({
+            'name': to_parse.stem,
+            'mets': mets,
+            'package': InformationPackages.details_from_mets_file(to_parse.joinpath('METS.xml'))
+        })
 
-class InformationPackage:
-    """Stores the vital facts and figures about a package."""
-    def __init__(self, path: str, details: PackageDetails, manifest: Manifest=None):
-        self._path = path
-        self._details = details
-        self._manifest = manifest if manifest else Manifest.from_directory(path)
-
-    @property
-    def path(self) -> str:
-        """Get the specification of the package."""
-        return self._path
-
-    @property
-    def details(self) -> PackageDetails:
-        """Get the package details."""
-        return self._details
-
-    @property
-    def manifest(self) -> Manifest:
-        """Return the package manifest."""
-        return self._manifest
+    @staticmethod
+    def validate(package_path: Path) -> Result:
+        if (not package_path.exists()):
+            raise FileNotFoundError(NO_PATH.format(package_path))
+        handler: PackageHandler = PackageHandler()
+        to_parse:Path = handler.prepare_package(package_path)
+        mets_path: Path = to_parse.joinpath('METS.xml')
+        if (not mets_path.is_file()):
+            raise ValueError('No METS file found in package')
+        return True
