@@ -1,6 +1,31 @@
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# flake8: noqa
+#
+# E-ARK Validation
+# Copyright (C) 2019
+# All rights reserved.
+#
+# Licensed to the E-ARK project under one
+# or more contributor license agreements. See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership. The E-ARK project licenses
+# this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License. You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
 
-from pydantic import BaseModel, Field
+from typing import Annotated, Optional
+from pydantic import BaseModel, Field, StringConstraints
 
 from enum import Enum
 import hashlib
@@ -16,7 +41,7 @@ class ChecksumAlg(str, Enum):
     SHA512 = 'SHA-512'
 
     @classmethod
-    def from_string(cls, value: str) -> 'ChecksumAlg':
+    def from_string(cls, value: str) -> Optional['ChecksumAlg']:
         search_value = value.upper() if hasattr(value, 'upper') else value
         for algorithm in ChecksumAlg:
             if (algorithm.value == search_value) or (algorithm.name == search_value) or (algorithm == value):
@@ -41,4 +66,4 @@ class ChecksumAlg(str, Enum):
 
 class Checksum(BaseModel):
     algorithm: ChecksumAlg = ChecksumAlg.SHA1
-    value: str = Field(default='', to_upper=True)
+    value: Annotated[ str, StringConstraints(to_upper=True) ] = ''
