@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# flake8: noqa
 #
 # E-ARK Validation
 # Copyright (C) 2019
@@ -23,20 +22,13 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-"""
-E-ARK : Information Package Validation
-        Information Package API model types
-"""
-# import models into model package
-from .checksum import Checksum, ChecksumAlg
-from .manifest import Manifest, ManifestEntry, SourceType
-from .validation_report import ValidationReport
-from .package_details import PackageDetails
-from .package_details import Representation
-from .validation_report import (
-        Level,
-        Severity,
-        StructureStatus,
-        StructResults,
-        Result
-)
+"""Utilities used across validation modules"""
+from pathlib import Path
+
+from .const import NO_PATH
+
+def get_path(path: Path | str, check_exists: bool=False) -> Path:
+    result: Path = Path(path) if isinstance(path, str) else path
+    if check_exists and not result.exists():
+        raise FileNotFoundError(NO_PATH.format(path))
+    return result
