@@ -22,11 +22,13 @@
 # specific language governing permissions and limitations
 # under the License.
 #
+from typing import List
 import unittest
 
 from importlib_resources import files
 
 from eark_validator import rules as SC
+from eark_validator.model.validation_report import Severity, Result
 import tests.resources.schematron as SCHEMATRON
 import tests.resources.xml as XML
 
@@ -72,20 +74,27 @@ class SchematronTest(unittest.TestCase):
 
     def test_validate_person(self):
         self._person_rules.validate(str(files(XML).joinpath(PERSON_XML)))
-        self.assertTrue(SC.TestReport.from_validation_report(self._person_rules._schematron.validation_report).is_valid)
+        self.assertTrue(_is_list_valid(SC.TestResults.from_validation_report(self._person_rules._schematron.validation_report)))
 
     def test_validate_invalid_person(self):
         self._person_rules.validate(str(files(XML).joinpath('invalid-person.xml')))
-        self.assertFalse(SC.TestReport.from_validation_report(self._person_rules._schematron.validation_report).is_valid)
+        self.assertFalse(_is_list_valid(SC.TestResults.from_validation_report(self._person_rules._schematron.validation_report)))
 
     def test_validate_mets(self):
         self._mets_one_def_rules.validate(METS_VALID_PATH)
-        self.assertTrue(SC.TestReport.from_validation_report(self._mets_one_def_rules._schematron.validation_report).is_valid)
+        self.assertTrue(_is_list_valid(SC.TestResults.from_validation_report(self._mets_one_def_rules._schematron.validation_report)))
 
     def test_validate_mets_no_root(self):
         self._mets_one_def_rules.validate(str(files(XML).joinpath('METS-no-root.xml')))
-        self.assertFalse(SC.TestReport.from_validation_report(self._mets_one_def_rules._schematron.validation_report).is_valid)
+        self.assertFalse(_is_list_valid(SC.TestResults.from_validation_report(self._mets_one_def_rules._schematron.validation_report)))
 
     def test_validate_mets_no_objid(self):
         self._mets_one_def_rules.validate(str(files(XML).joinpath('METS-no-objid.xml')))
-        self.assertFalse(SC.TestReport.from_validation_report(self._mets_one_def_rules._schematron.validation_report).is_valid)
+        self.assertFalse(_is_list_valid(SC.TestResults.from_validation_report(self._mets_one_def_rules._schematron.validation_report)))
+
+def _is_list_valid(to_test: List[Result]) -> bool:
+    for result in to_test:
+        if result.severity == Severity.Error:
+            return False
+    return True
+    # return len(list(filter(lambda a: a == Severity.Error, to_test))) < 1
